requirements.txt
streamlit==1.41.1
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langchain-huggingface==0.0.3
sentence-transformers==3.3.1
chromadb==0.5.20
pydantic==2.10.4
python-dotenv==1.0.1
pysqlite3-binary==0.5.4
